requirements.txt
alembic==1.8.1
attrs==22.1.0
bidict==0.22.0
click==8.1.3
dnspython==1.16.0
eventlet==0.30.2
Flask==2.2.2
Flask-Assets==2.0
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
htmlmin==0.1.12
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsmin==3.0.1
jsonlines==3.1.0
lesscpy==0.15.0
Mako==1.2.2
MarkupSafe==2.1.1
ply==3.11
psycopg2-binary==2.9.3
python-dotenv==0.20.0
python-engineio==4.3.4
python-socketio==5.7.1
random-username==1.0.2
rcssmin==1.1.1
simple-websocket==0.8.0
six==1.16.0
SQLAlchemy==1.4.40
webassets==2.0
Werkzeug==2.2.2
wsproto==1.1.0
xxhash==3.0.0
zipp==3.8.1
